requirements.txt
asn1crypto==0.24.0
cryptography==2.1.4
Django==1.11.20
djangorestframework==3.9.4
enum34==1.1.6
idna==2.6
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pycrypto==2.6.1
pygobject==3.26.1
pyserial==3.4
pytz==2019.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
unity-lens-photos==1.0
virtualenv==16.5.0
